extract ranked full solve table into component
