+++ frontend/src/pages/Puzzle.tsx
@@ -1,6 +1,7 @@
 import CategoryQuerySelector from "@/components/category-query-selector";
 import Header from "@/components/header"
 import ProgramIcon from "@/components/icon/program-icon";
+import RankedFullSolveTable from "@/components/ranked-full-solve-table";
 import RecordHistoryChart from "@/components/record-history-chart";
 import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from "@/components/ui/table";
 import { getPuzzleInfo, getPuzzleSolves, type RankedFullSolve, type Puzzle, getRecordHistory, type FullSolve } from "@/lib/backend";
@@ -9,7 +10,6 @@ import { useEffect, useMemo, useState } from "react"
 import { Link, useNavigate, useSearchParams } from "react-router-dom"
 
 function Puzzle() {
-    const navigate = useNavigate()
 
     const [searchParams] = useSearchParams()
     const id = Number(searchParams.get("id")) || 1
@@ -41,43 +41,10 @@ function Puzzle() {
             {searchParams.get("history") === "true" ? (
                 <RecordHistoryChart history={history}/>
             )
-            : <></>}
-
-            <Table>
-                <TableHeader>
-                    <TableRow >
-                        <TableHead>Rank</TableHead>
-                        <TableHead>Solver</TableHead>
-                        <TableHead>{searchParams.get("event") === "fmc" || searchParams.get("event") === "fmcca" ? "Move count" : "Time"}</TableHead>
-                        <TableHead>Date</TableHead>
-                        <TableHead>Program</TableHead>
-                    </TableRow>
-                </TableHeader>
-                {solves &&
-                <TableBody>
-                    {solves.length > 0 ? solves.map((s) => (
-                        <TableRow onClick={() => navigate(`/solve?id=${s.solve.id}`)}>
-                            <TableCell>{s.rank}</TableCell>
-                            <TableCell>
-                                <Link to={`/solver?id=${s.solve.solver.id}`} onClick={(e) => e.stopPropagation()}>
-                                    {s.solve.solver.name}
-                                </Link>
-                            </TableCell>
-                            <TableCell>
-                                {searchParams.get("event") === "fmc" || searchParams.get("event") === "fmcca" ?
-                                s.solve.move_count && s.solve.move_count :
-                                s.solve.speed_cs && html_render_time(s.solve.speed_cs)}
-                            </TableCell>
-                            <TableCell>{html_render_date(s.solve.solve_date)}</TableCell>
-                            <TableCell className="inline-flex items-center"><ProgramIcon abbr={s.solve.program.abbr}/>&nbsp;{s.solve.program.abbr}</TableCell>
-                        </TableRow>
-                    ))
-                    :
-                        <p className="pt-4">There are no solves matching the search.</p>
-                    }
-                </TableBody>
+            : <RankedFullSolveTable RankedSolves={solves} isFmc={searchParams.get("event") === "fmc" || searchParams.get("event") === "fmcca"}/>
             }
-            </Table>
+
+
         </>
     )
 }
